refactor: 데이터 타입 최적화
- 가격 타입 변경에 따른 Service, DTO 최적화
- PortOne 요청/응답 데이터 타입 맞춤

Refs: #88

+++ src/main/java/com/sparta/cream/dto/portone/PortOnePaymentResponse.java
@@ -16,7 +16,7 @@ public class PortOnePaymentResponse {
     @NoArgsConstructor
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Amount {
-        private Long total;
+        private int total;
     }
 
     @Getter
